pyproject.toml
[project]
name = "google-pulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "fastmcp>=3.2.0",
    "google-adk>=1.28.0",
    "google-api-python-client>=2.193.0",
    "google-auth-oauthlib>=1.3.1",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
